Updates vmci_client C library for DLL generation esx_vmdkcmd.go uses CGO for making vmci_client calls to communicate with the VMDKOps service on ESXi. By default, CGO supports gcc, which creates compatibility issues on Windows, because, the vmci_client implementation depends on libraries provided by Microsoft Build Tools. This commit enables exporting of the vmci_client library as a Windows DLL, which can then be used with the regular MinGW-w64 gcc on Windows for making calls via CGO.
* Extracted header file from vmci_client.c
* Updated vmci_client impl to support Winsock
* Added vmci_client.def for DLL export
